requirements.txt
amqp==5.1.1
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
ifaddr==0.1.7
kombu==5.2.4
py-eureka-client==0.10.5
pymongo==3.12.3
pytz==2022.1
sqlparse==0.2.4
tzdata==2022.1
vine==5.0.0
yosun==0.6.3
